AI-Documentation/requirements.txt
tree-sitter==0.24.0
tree-sitter-python==0.23.6
llama-cpp-python==0.3.9
